# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Web API Express",
    "description": "My WebApi using Node",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Players",
      "description": "API de jugadores en la aplicación"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/getPlayers": {
      "get": {
        "tags": ["Players"],
        "summary": "Obtiene todos los jugadores",
        "description": "Regresa todos los jugadores",
        "operationId": "getPlayers",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        }
      }
    },
    "/api/getPlayer/{id}": {
      "get": {
        "tags": ["Players"],
        "summary": "Encuentra un jugador por id",
        "description": "Regresa el jugador con el id solicitado",
        "operationId": "getPlayer",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id del jugador a actualizar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        }
      }
    },
    "/api/addPlayer": {
      "post": {
        "tags": ["Players"],
        "summary": "Agrega un nuevo jugador",
        "description": "Agrega un jugador a la base de datos",
        "operationId": "addPlayer",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Nuevo jugador",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          },
          "405": {
            "description": "Datos inválidos"
          }
        }
      }
    },
    "/api/updatePlayer/{id}": {
      "put": {
        "tags": ["Players"],
        "summary": "Actualiza un nuevo jugador",
        "description": "Actualiza un jugador a la base de datos",
        "operationId": "updatePlayer",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id del jugador a actualizar",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Datos nuevos jugador",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          },
          "400": {
            "description": "Jugador Inválido"
          },
          "404": {
            "description": "Jugador no encontrado"
          }
        }
      }
    },
    "/api/deletePlayer/{id}": {
      "delete": {
        "tags": ["Players"],
        "summary": "Borra un jugador",
        "description": "Borra un jugador de la base de datos",
        "operationId": "deletePlayer",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id del jugador a borrar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          },
          "400": {
            "description": "Jugador Inválido"
          },
          "404": {
            "description": "Jugador no encontrado"
          }
        }
      }
    },
    "/api/json/Terniumeq2": {
      "get": {
        "tags": ["Quiz"],
        "summary": "Encuentra palabra e imagen",
        "description": "Regresa la palabra y la imagen",
        "operationId": "getWord",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Quiz"
            }
          }
        }
      }
    },
    "/api/json/loginPlayer/{email}": {
      "get": {
        "tags": ["Login"],
        "summary": "Login de usuario",
        "description": "Revisa que los datos sean correctos",
        "operationId": "login",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "Email del usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        }
      }
    },
    "/api/json/getStatistics": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Obtiene todas las estadísticas",
        "description": "Regresa todas las estadísticas",
        "operationId": "getStatistics",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Statistics"
            }
          }
        }
      }
    }, 
    "/api/json/statistics/{name}": {
      "get": {
        "tags": ["Estadisticas"],
        "summary": "Muestra historial de estadisticas del jugdor",
        "description": "regresa todas las estadisticas del jugador",
        "operationId": "getStatisticsName",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "description": "Name del usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Estadisticas"
            }
          }
        }
      }
    },
    "/api/json/getStatisticsID/{id}": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Obtiene todas las estadísticas",
        "description": "Regresa todas las estadísticas",
        "operationId": "getStatisticsID",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id del usuario",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Statistics"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Player": {
      "type": "object",
      "properties": {
        "Nombre": {
          "type": "string"
        },
        "Apellido": {
          "type": "string"
        },
        "Email": {
          "type": "string",
          "format": "int64"
        },
        "Password": {
          "type": "string"
        },
        "EsAdministrador": {
          "type": "boolean"
        }
      }
    },
    "Quiz": {
      "type": "object",
      "properties": {
        "answer": {
          "type": "string",
          "format": "int64"
        },
        "questionImage": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "Email": {
          "type": "string",
          "format": "int64"
        },
        "Password": {
          "type": "string"
        },
        "EsAdministrador": {
          "type": "boolean"
        }
      }
    },
    "Statistics": {
      "type": "object",
      "properties":{
        "playerID": {
          "type": "integer"
        },
        "statisticID": {
          "type": "integer"
        },
        "timeDate": {
          "type": "string"
        },
        "constructionNumber": {
          "type": "integer"
        },
        "totalMoney": {
          "type": "number"
        },
        "totalSteel": {
          "type": "number"
        },
        "totalContract": {
          "type": "integer"
        },
        "playerExperience": {
          "type": "integer"
        },
        "totalIron":{
          "type": "integer"
        },
        "totalUnpackageSteel":{
          "type": "integer"
        },
        "ironToSteelTransform": {
          "type": "number"
        }
      }
    }, 
    "Estadisticas": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "timeDate": {
          "type": "string"
        },
        "constructionNumber": {
          "type": "integer",
          "format": "int64"
        },
        "totalMoney": {
          "type": "integer",
          "format": "int64"
        },
        "totalSteel": {
          "type": "integer",
          "format": "int64"
        },
        "totalContract": {
          "type": "integer",
          "format": "int64"
        },
        "playerExperience": {
          "type": "integer",
          "format": "int64"
        },
        "ironToSteelTransform": {
          "type": "integer",
          "format": "int64"
        }
      }
    }
  }
}